+++ libraries/myqee/core/classes/MyQEE/Database/Driver/Mongo.class.php
@@ -192,6 +192,8 @@ class MyQEE_Database_Driver_Mongo extends Database_Driver
                 }
                 catch ( Exception $e )
                 {
+                    if (IS_DEBUG)Core::debug()->error(($username?$username.'@':'').$hostname.':'.$port,'connect mongodb server error');
+
                     if (2==$i && !in_array($hostname, $error_host))
                     {
                         $error_host[] = $hostname;
@@ -950,7 +952,14 @@ class MyQEE_Database_Driver_Mongo extends Database_Driver
         elseif ( isset($tmp_query[$now_logic]) )
         {
             // 如果有 $and $or 条件，则加入
-            $tmp_query[$now_logic][] = $tmp_option;
+            if ( is_array($tmp_option) || !$column )
+            {
+                $tmp_query[$now_logic][] = $tmp_option;
+            }
+            else
+            {
+                $tmp_query[$now_logic][] = array($column=>$tmp_option);
+            }
         }
         else if ($column)
         {
@@ -1000,20 +1009,14 @@ class MyQEE_Database_Driver_Mongo extends Database_Driver
                     }
 
                 }
-                elseif ( is_array($tmp_option) || is_array($tmp_query[$column]) )
+                else
                 {
-                    # 有一个是数组
                     $tmp_query['$and'] = array
                     (
                         array( $column => $tmp_query[$column] ),
                         array( $column => $tmp_option ),
                     );
                     unset($tmp_query[$column]);
-                }
-                else
-                {
-                    # 都是字符串
-                    $tmp_query[$column] = $tmp_option;
                 }
             }
             else