+++ libraries/myqee/core/classes/MyQEE/Database/Driver/MySQLI.class.php
@@ -205,6 +205,8 @@ class MyQEE_Database_Driver_MySQLI extends Database_Driver
                 }
                 catch ( Exception $e )
                 {
+                    if (IS_DEBUG)Core::debug()->error($username.'@'.$hostname.':'.$port,'connect mysqli server error');
+
                     $last_error = $e;
                     if (2===$e->getCode() && preg_match('#(Unknown database|Access denied for user)#i', $e->getMessage()))
                     {