+++ libraries/myqee/core/classes/MyQEE/Database/Driver/SQLite.class.php
@@ -584,7 +584,7 @@ class MyQEE_Database_Driver_SQLite extends Database_Driver
         // Start a selection query
         $query = 'SELECT ';
 
-        if ( $builder['distinct'] === true )
+        if ( $builder['distinct'] )
         {
             // Select only unique results
             $query .= 'DISTINCT ';