+++ libraries/myqee/core/classes/MyQEE/Database/QueryBuilder.class.php
@@ -114,7 +114,20 @@ class MyQEE_Database_QueryBuilder
     }
 
     /**
-     * Adds or overwrites values. Multiple value sets can be added.
+     * 加入多条数据
+     *
+     *     // 例1
+     *     $v1 = array('k1'=>1,'k2'=>1);
+     *     $v2 = array('k1'=>2,'k2'=>1);
+     *     $v3 = array('k1'=>3,'k2'=>1);
+     *     $db->values($v1,$v2,$v3);        //加入3行数据
+     *
+     *     // 例2
+     *     $values = array();
+     *     $values[] = array('k1'=>1,'k2'=>1);
+     *     $values[] = array('k1'=>2,'k2'=>1);
+     *     $values[] = array('k1'=>3,'k2'=>1);
+     *     $db->values($values);            //加入3行数据,等同上面的效果
      *
      * @param   array   values list
      * @param   ...
@@ -122,13 +135,15 @@ class MyQEE_Database_QueryBuilder
      */
     public function values(array $values)
     {
-        if ( ! is_array($this->_builder['values']) )
+        if ( is_array($values) && isset($values[0]) && is_array($values[0]) )
         {
-            throw new Exception('INSERT INTO ... SELECT statements cannot be combined with INSERT INTO ... VALUES');
+            // 多行数据
+            // $values = $values;
         }
-
-        // Get all of the passed values
-        $values = func_get_args();
+        else
+        {
+            $values = func_get_args();
+        }
 
         $this->_builder['values'] = array_merge($this->_builder['values'], $values);