+++ libraries/myqee/core/classes/MyQEE/File.class.php
@@ -925,7 +925,11 @@ class MyQEE_File
     protected static function sync_mode()
     {
         static $mode = null;
-        if ( null === $mode ) $mode = Core::config('core.file_sync_mode');
+        if ( null === $mode )
+        {
+            $mode = Core::config('core.file_sync_mode');
+            if (null===$mode)$mode = 'default';
+        }
 
         return $mode;
     }